+++ bin/benchmark/Cargo.toml
@@ -23,18 +23,18 @@ tokio = { version = "1", features = ["full"] }
 futures = "0.3"
 rand = "0.8"
 indicatif = "=0.17.1"
+bb8-redis = { version = "0.14", optional = true }
 
 [dependencies.fred]
-#path = "../.."
-path = "/fred"
+path = "../.."
+#path = "/fred"
 features = ["replicas", "unix-sockets"]
-#features = ["replicas"]
-#version = "7.1.1"
 default-features = false
 
 [features]
 default = []
 assert-expected = []
+redis-rs = ["bb8-redis"]
 enable-rustls = ["fred/enable-rustls"]
 enable-native-tls = ["fred/enable-native-tls"]
 debug-ids = ["fred/debug-ids"]
@@ -42,3 +42,5 @@ stdout-tracing = ["fred/partial-tracing"]
 partial-tracing = ["fred/partial-tracing"]
 full-tracing = ["fred/full-tracing"]
 blocking-encoding = ["fred/blocking-encoding"]
+
+[workspace]
