+++ bin/benchmark/src/utils.rs
@@ -1,6 +1,12 @@
-use std::sync::atomic::AtomicUsize;
-use std::sync::atomic::Ordering;
-use std::sync::Arc;
+use rand::{self, distributions::Alphanumeric, Rng};
+use std::{
+  env,
+  error::Error,
+  sync::{
+    atomic::{AtomicUsize, Ordering},
+    Arc,
+  },
+};
 
 pub fn incr_atomic(size: &Arc<AtomicUsize>) -> usize {
   size.fetch_add(1, Ordering::AcqRel).saturating_add(1)
@@ -16,4 +22,24 @@ pub fn read_atomic(size: &Arc<AtomicUsize>) -> usize {
 
 pub fn set_atomic(size: &Arc<AtomicUsize>, val: usize) -> usize {
   size.swap(val, Ordering::SeqCst)
+}
+
+pub fn read_auth_env() -> (Option<String>, Option<String>) {
+  let username = env::var_os("REDIS_USERNAME").and_then(|s| s.into_string().ok());
+  let password = env::var_os("REDIS_PASSWORD").and_then(|s| s.into_string().ok());
+
+  (username, password)
+}
+
+pub fn random_string(len: usize) -> String {
+  rand::thread_rng()
+    .sample_iter(&Alphanumeric)
+    .take(len)
+    .map(char::from)
+    .collect()
+}
+
+pub fn crash(error: impl Error) {
+  println!("{:?}", error);
+  std::process::exit(1);
 }
